move sql code
